requirements.txt
# local package
-e .

# external requirements
click
# Sphinx
coverage
# awscli
flake8
python-dotenv>=0.5.1
numpy==1.23.4
pandas==1.5.1
scikit-learn==1.1.3
matplotlib==3.6.2
seaborn==0.12.1
ipython==8.6.0
jupyterlab==3.5.0
tabulate==0.9.0
statsmodels==0.13.5
xgboost==1.7.1
tensorflow-cpu==2.11.0

# backwards compatibility
pathlib2
